# d2021/docs/swagger.yaml
basePath: /path
definitions:
  d2021.CoursesConv:
    properties:
      title:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8001
info:
  contact: {}
  description: API Server for HRM system
  title: HRM API
  version: "1.0"
paths:
  /employee/course:
    post:
      consumes:
      - application/json
      description: Start taking the course assigned to the employee
      operationId: post-course-to-start
      parameters:
      - description: Employee course
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/d2021.CoursesConv'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: PostEmployeeCourses
      tags:
      - Course
  /finished:
    get:
      consumes:
      - application/json
      description: Get finished by employee courses
      operationId: get-finished-course
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d2021.CoursesConv'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetFinishedCourses
      tags:
      - Course
  /progress:
    get:
      consumes:
      - application/json
      description: Get started by employee courses
      operationId: get-started-course
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d2021.CoursesConv'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetStartedCourses
      tags:
      - Course
  /suggested:
    get:
      consumes:
      - application/json
      description: Get suggested courses for employee
      operationId: get-suggested-course
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d2021.CoursesConv'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetCourse
      tags:
      - Course
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"